pyproject.toml
[tool.poetry]
name = "zaks-clouds-manager"
version = "0.1.0"
description = ""
authors = ["Matheus Albino Brunhara <matheusalb3213@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.29.3"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
